Change example AD account from mssql to mssqlAD
I felt the name used as the example Active Directory account "mssql" added confusion because that is also the Linux account under which SQL Server runs. This led to some confusion in later Linux commands where the ownership of the keytab file is being set with chown/chmod commands. I think changing the example AD account to "mssqlAD" helps avoid that confusion, so that is what I did. I also changed the kinit/kvno section to use "mssqlAD@CONTOSO.COM" instead of "user@CONTOSO.COM" to be consistent with using the actual example account name instead of just "user".